Move flanking rule under combat rules
