+++ Assets/Scripts/Combat/Rules.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 761b48fa4625bed45b3b20081e4d5e39
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
